fix: Adds CUO settings support and adds more robust 7.0.9 support
